Fix #17: remove a set from the routine editor
The routine editor could add sets but never remove one — there was no
affordance at all. Mirror the active-workout screen's existing UX rather
than invent a second one: swipe the set row left to reveal a Delete
panel, with the parent keeping at most one row open.

To keep the two identical, the swipe mechanics move out of SetRow into a
shared SwipeToDelete wrapper (pan gate, 220ms easing, -72px open offset,
76px panel — all unchanged) which both screens now use. SetRow renders
the same as before; only where the animation lives has moved.

Removing the last set of an exercise is allowed, matching deleteSet on
the workout screen; "+ Add Set" is right there to undo it.
